requirements.txt
torch
kafka-python
jsonschema
numpy
